src/app/components/layout/layout.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { LayoutComponent } from './layout.component';
import { RouterTestingModule } from '@angular/router/testing';
import { AuthService } from 'src/app/services/auth/auth.service';
import { StorageService } from 'src/app/services/storage/storage.service';

describe('LayoutComponent', () => {
  let component: LayoutComponent;
  let fixture: ComponentFixture<LayoutComponent>;
  let authService: jasmine.SpyObj<AuthService>;
  let storageService: jasmine.SpyObj<StorageService>;

  beforeEach(async () => {
    const authServiceSpy = jasmine.createSpyObj('AuthService', ['canAuthenticate']);
    const storageServiceSpy = jasmine.createSpyObj('StorageService', ['getToken']);

    await TestBed.configureTestingModule({
      declarations: [LayoutComponent],
      imports: [RouterTestingModule],
      providers: [
        { provide: AuthService, useValue: authServiceSpy },
        { provide: StorageService, useValue: storageServiceSpy }
      ]
    }).compileComponents();

    authService = TestBed.inject(AuthService) as jasmine.SpyObj<AuthService>;
    storageService = TestBed.inject(StorageService) as jasmine.SpyObj<StorageService>;
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(LayoutComponent);
    component = fixture.componentInstance;
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should call authService canAuthenticate if token exists on initialization', () => {
    storageService.getToken.and.returnValue('exampleToken');
    component.ngOnInit();
    expect(authService.canAuthenticate).toHaveBeenCalled();
  });

  it('should navigate to not-found route if token does not exist on initialization', () => {
    storageService.getToken.and.returnValue(null);
    const navigateSpy = spyOn((<any>component).router, 'navigate');
    component.ngOnInit();
    expect(navigateSpy).toHaveBeenCalledWith(['not-found']);
  });
});
